+++ noxfile.py
@@ -1,4 +1,5 @@
 """Nox sessions."""
+
 import os
 import platform
 import tempfile
@@ -8,7 +9,7 @@ import nox
 from nox.sessions import Session
 
 nox.options.sessions = ["tests", "mypy"]
-python_versions = ["3.8", "3.9", "3.10", "3.11"]
+python_versions = ["3.11", "3.12"]
 
 
 def install_with_constraints(session: Session, *args: str, **kwargs: Any) -> None:
@@ -81,7 +82,7 @@ def mypy(session: Session) -> None:
     session.run("inv", "mypy")
 
 
-@nox.session(python="3.10")
+@nox.session(python="3.11")
 def safety(session: Session) -> None:
     """Scan dependencies for insecure packages."""
     install_with_constraints(session, "invoke", "safety")