+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2024-03-23
+
 ### Added
 - Python `3.12` is now fully tested and supported in CI/CD.
 - New module `translator_typing`. Useful especially users who which to extend the base `Translator` implementation.
@@ -227,7 +229,8 @@ cookiecutter template.
 - Switch to relative imports.
 - Fix some intersphinx issues.
 
-[Unreleased]: https://github.com/rsundqvist/id-translation/compare/v0.7.1...HEAD
+[Unreleased]: https://github.com/rsundqvist/id-translation/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/rsundqvist/id-translator/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/rsundqvist/id-translator/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/rsundqvist/id-translator/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/rsundqvist/id-translator/compare/v0.5.1...v0.6.0